Move test credentials
